refactor(ui): refactor Buttons
