--- ui/design/src/components/Buttons/index.ts
@@ -1,17 +1,0 @@
-import ProfileAvatarButton from './avatar-buttons/profile-avatar-button';
-import Button from './default-buttons/button';
-import { DuplexButton } from './default-buttons/duplex-button';
-import PlainButton from './default-buttons/plain-button';
-import IconButton from './icon-buttons/icon-button';
-import IconLink from './icon-buttons/icon-link';
-import VoteIconButton from './icon-buttons/vote-icon-button';
-
-export {
-  Button,
-  DuplexButton,
-  IconButton,
-  IconLink,
-  ProfileAvatarButton,
-  VoteIconButton,
-  PlainButton,
-};
