+++ ui/design/src/components/Cards/form-cards/box-form-card.tsx
@@ -1,7 +1,7 @@
 import { Box, FormField, Text, TextArea, TextInput } from 'grommet';
 import * as React from 'react';
 import { isMobile, isMobileOnly } from 'react-device-detect';
-import { Button } from '../../Buttons/index';
+import Button from '../../Button';
 import { Icon } from '../../Icon/index';
 import { StyledLayer } from '../../Modals/common/styled-modal';
 import { StyledImageInput } from '../../Popovers/image/styled-form-image-popover';
