+++ ui/design/src/components/Cards/form-cards/ens-form-card.tsx
@@ -1,7 +1,7 @@
 import { Box, RadioButton, Text } from 'grommet';
 import * as React from 'react';
 import styled from 'styled-components';
-import { Button } from '../../Buttons/index';
+import Button from '../../Button';
 import { Icon } from '../../Icon/index';
 import { StyledLayer } from '../../Modals/common/styled-modal';
 import Spinner from '../../Spinner';
