+++ ui/design/src/components/Cards/profile-cards/profile-card-fields/profile-card-cover-image.tsx
@@ -1,6 +1,6 @@
 import { Box, Text } from 'grommet';
 import * as React from 'react';
-import { IconButton } from '../../../Buttons/index';
+import IconButton from '../../../IconButton';
 import { Icon } from '../../../Icon';
 import { AppIcon } from '../../../Icon/index';
 import { SelectPopover } from '../../../Popovers/index';
