+++ ui/design/src/components/Cards/tag-cards/tag-detail-card.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { ITag } from '../widget-cards/trending-widget-card';
 import { SubtitleTextIcon } from '../../TextIcon';
 import { Box, Text } from 'grommet';
-import { DuplexButton, IconButton } from '../../Buttons';
+import IconButton from '../../IconButton';
+import DuplexButton from '../../DuplexButton';
 import { IMentionData } from '../../Charts/area-chart';
 import { ResponsiveLineChart } from '../../Charts';
 import { Icon } from '../../Icon';
