+++ ui/design/src/components/Input/link-input/styled-link-input.ts
@@ -1,6 +1,5 @@
 import { TextInput, FormField } from 'grommet';
 import styled from 'styled-components';
-// import { Button } from '../../Buttons/index';
 
 const StyledLinkIconDiv = styled.div`
   height: 2rem;
