+++ ui/design/src/components/Popovers/app-menu/app-menu-popover.tsx
@@ -1,7 +1,7 @@
 import { Box, Text } from 'grommet';
 import * as React from 'react';
 import { IMenuItem } from '@akashaproject/ui-awf-typings/lib/app-loader';
-import { IconLink } from '../../Buttons';
+import IconLink from '../../IconLink';
 import { Icon } from '../../Icon/index';
 import {
   StyledDrop,
