+++ ui/design/src/components/TopBar/index.tsx
@@ -18,7 +18,7 @@ import {
   IconDiv,
 } from './styled-topbar';
 import { isMobileOnly } from 'react-device-detect';
-import { Button } from '../Buttons';
+import Button from '../Button';
 import { IProfileData } from '../Cards/profile-cards/profile-widget-card';
 import { ProfileMenu } from './profile-menu';
