+++ ui/design/src/components/VoteIconButton/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { Icon } from '../../Icon';
-import StyledIconButton from './styled-icon-button';
+import { Icon } from '../Icon';
+import StyledIconButton from '../IconButton/styled-icon-button';
 
 export interface IVoteIconProps {
   className?: string;
